requirements.txt
freeze
Django==1.10.1
djangorestframework==3.4.7
django-rest-swagger==2.0.7
